Move store.config to /src and make it TS
